refactoring: genericize as much as possible
It is ideal to minimize the number of functions that require
implementation for each `FiscalCalendar`. The goal is to limit this to
the following list for the given calendar `MY_C<:FiscalCalendar`:
* `lastday(::AccountingPeriod{MY_C,FiscalYear})`
* `periodofFY(::AccountingPeriod{MY_C,D}) where D`
* `periodofQ(::AccountingPeriod{MY_C,D}) where D`
* `weekofFY(::AccountingPeriod{MY_C,FiscalWeek}) where D`
* `weekofQ(::AccountingPeriod{MY_C,FiscalWeek})`
* `weekofP(::AccountingPeriod{MY_C,FiscalWeek})`

Interestingly, `firstday` can be completely genericized, even in spite
of the difficulty of its implementation with respect to `FiscalPeriod`s.
The idea is to use the periodicity of the durations of `FiscalPeriod`s
as a type parameter to the `FiscalCalendar`. For example, accountants
might speak of 4-4-5, 4-5-4, of 5-4-4 calendars, which, are really
periodic week-lengths of the periods of each quarter for non-leap years.
In leap years, it can be safely assumed that the "extra week" will be
added to the last 4-length period. Additionally, there might be a 4th
possibly value for this type parameter: a 13-period calendar. In this
one, every period is 4 weeks, other than the last one in leap years
only. With this implementation, it should be possible to exclude
`firstday` entirely from the list above.